refactor: move the check to the solution file
